% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_trackers.R
\name{add_trackers}
\alias{add_trackers}
\title{Add tracker columns}
\usage{
add_trackers(x)
}
\arguments{
\item{x}{A data frame of focal observations}
}
\value{
Adds three columns: si_ei_tracker, bc_tracker, da_pa_tracker
}
\description{
Add tracker columns
}
