% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_da_ei.R
\name{check_da_ei}
\alias{check_da_ei}
\title{Check for da's without an ei}
\usage{
check_da_ei(x, offset = 1)
}
\arguments{
\item{x}{A data of focal observations}

\item{offset}{Number of seconds of difference to check for si/ei code.
Default = 1}
}
\description{
Currently needs these exact columns: "da_check", at_sec_since_mdn", "focal_animal", "action_partner", "date"
}
\details{
da_check needs to only be an empty character column.
}
