% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing_start_end.R
\name{check_missing_start_end}
\alias{check_missing_start_end}
\title{Check for missing start/end codes}
\usage{
check_missing_start_end(x, start_codes = "sg", end_codes = "eg")
}
\arguments{
\item{x}{A data frame of focal observations}

\item{start_codes}{A character vector of start codes that indicate a behaviour has started (e.g. c("ap", "at")).}

\item{end_codes}{A character vector of end codes that indicate a behaviour has finished (e.g. c("de", do)).}
}
\value{
Returns the focal data with an added "check" column. Look for "check next"
which indicates where there two consecutive start or end codes.
}
\description{
Check for missing start/end codes
}
