% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_agg_action_intensity.R
\name{class_agg_action_intensity}
\alias{class_agg_action_intensity}
\title{Classify aggression intensity}
\usage{
class_agg_action_intensity(x, agg.low, agg.high, ids)
}
\arguments{
\item{x}{A data frame of focal observations}

\item{agg.low}{A character vector of low intensity aggression codes}

\item{agg.high}{A character vector of high intensity aggression codes}

\item{ids}{A character vector of aggression partner IDs}
}
\value{
A data frame with two new columns: for agg given and rec
}
\description{
This is a helper function for classify_da_id(). Will classify individual aggressions as high or low intensity.
}
