% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_da_id.R
\name{class_da_id}
\alias{class_da_id}
\title{Classify da pa id}
\usage{
class_da_id(x, tag = NULL)
}
\arguments{
\item{x}{A data frame of focal observations}

\item{tag}{A character string that denotes the class of aggression partner. will be added to the new column name. e.g. "mm" or "mf".}
}
\value{
A data frame with two new columns.
}
\description{
This is a helper function for classify_da_id(). Requires the output of class_agg_action_intensity().
Checks if an aggression inside a da_pa_id was given or received at least once. Prioritises high intensity over low.
}
