% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_da_id.R
\name{classify_da_id}
\alias{classify_da_id}
\title{Classify da_pa_ids}
\usage{
classify_da_id(x, agg.low, agg.high, id, tag = NULL)
}
\arguments{
\item{x}{A data frame of focal observations}

\item{agg.low}{A character vector of low intensity aggression codes}

\item{agg.high}{A character vector of high intensity aggression codes}

\item{id}{A character vector of aggression partner IDs}

\item{tag}{A character string that denotes the class of aggression partner. will be added to the new column name. e.g. "mm" or "mf".}
}
\value{
A data frame with two new columns.
}
\description{
Checks if an aggression inside a da_pa_id was given or received at least once. Prioritises high intensity over low.
}
