% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dur_matrix.R
\name{dur_matrix}
\alias{dur_matrix}
\title{Create a duration matrix}
\usage{
dur_matrix(action_totals, matrix_rows, matrix_cols)
}
\arguments{
\item{action_totals}{A data frame with durations}

\item{m_rows}{A character vector of IDs to go in matrix rows}

\item{m_cols}{A character vector of IDs to go in matrix columns}
}
\value{
A matrix with total duratioin per dyads (in seconds)
}
\description{
Convert a data frame, usually created by duration_calculator(),
with durations for a behaviour to a matrix.
}
