% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration_calculator.R
\name{duration_calculator}
\alias{duration_calculator}
\title{Calculate the duration of a behaviour}
\usage{
duration_calculator(d_focal, start_codes, end_codes, focals, partners = NULL)
}
\arguments{
\item{d_focal}{data frame of focal observations.
Currently needs all these columns with exact same names:
"date", "protocol_start_time", "focal_animal", "observer", "action_time",
"actor", "action", "receiver", "action_partner", "at_sec_since_mdn"}

\item{start_codes}{a character vector of action codes
that indicate when a behaviour starts.}

\item{end_codes}{a character vector of action codes
that indicate when a behaviour ends.}

\item{focals}{a character vector of focal animals}

\item{partners}{OPTIONAL character vector of action partners}
}
\value{
adds two columns to d_focal. One with the total duration in seconds
(duration_s), and another with the total duration in hh:mm:ss (duration)
}
\description{
Calculates the duration of behaviours in a focal protocol.
}
