% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyad_obs_total.R
\name{dyad_obs_total}
\alias{dyad_obs_total}
\title{function to add obs time of focal and participant}
\usage{
dyad_obs_total(obs_time, fcl, prt)
}
\arguments{
\item{obs_time}{A data frame of observation time, 
created by obs_time_calculator()}

\item{fcl}{Observation time of focal animal in "hh:mm:ss"}

\item{prt}{Observation time of partner in "hh:mm:ss"}
}
\value{
The sum of fcl and prt in seconds
}
\description{
function to add obs time of focal and participant
}
