% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_matrix.R
\name{freq_matrix}
\alias{freq_matrix}
\title{Create a frequency matrix}
\usage{
freq_matrix(action_df, matrix_rows, matrix_cols)
}
\arguments{
\item{action_df}{A data frame with all occurences of a single behaviour.
Usually created by duration_calculator().}

\item{m_rows}{A character vector of IDs to go in matrix rows}

\item{m_cols}{A character vector of IDs to go in matrix columns}
}
\value{
A matrix with frequencies per dyad
}
\description{
Create a matrix with the frequency that a certain behaviour occurs
for a particular dyad (e.g. proximity, body contact or grooming bouts)
}
