% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_obs_time_s.R
\name{get_obs_time_s}
\alias{get_obs_time_s}
\title{Get obs time in secs}
\usage{
get_obs_time_s(obs_time, id)
}
\arguments{
\item{obs_time}{data frame of obs time created by obs_time_calculator()}

\item{id}{character vector of animal IDs}
}
\value{
will return vector of single numeric value in seconds
}
\description{
Function to extract obstime in seconds of particular focal
}
