% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_si_rev.R
\name{highlight_si_rev}
\alias{highlight_si_rev}
\title{Highlight da-pa in reverse}
\usage{
highlight_si_rev(x, action_codes)
}
\arguments{
\item{x}{A data frame of focal observations}

\item{action_codes}{A character vector of action codes}
}
\description{
function to highlight beggining to end interaction of da and pa
action_codes: if previous action is present in action_codes then it will keep
highlighting. Highlighter stops if code is not present.
}
