% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_sibc.R
\name{highlight_sibc}
\alias{highlight_sibc}
\title{Highlight si/ei of affiliative behaviours}
\usage{
highlight_sibc(x, offset = 1)
}
\arguments{
\item{x}{A data frame of focal observations}

\item{offset}{Number of seconds of difference to check for si/ei code.
Default = 1}
}
\description{
Currently needs these exact columns: "si_ei_tracker", "bc_tracker",
"at_sec_since_mdn", "focal_animal", "action_partner", "date"
}
\details{
Tracker columns can be added with add_trackers()

Checks if si/ei codes belongs to sb/sp/tb, and highlights if it does.
}
