% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_nest.R
\name{map_nest}
\alias{map_nest}
\title{Map function to list-column}
\usage{
map_nest(x, ..., f)
}
\arguments{
\item{x}{A data frame}

\item{...}{Variables to group_by() (comma-separated and unquoted)}

\item{f}{A function passed on to map()}
}
\value{
Returns unested x after mapping function
}
\description{
1. Nest data by grouping variable(s)
2. Map function to nested list column
3. Unest data frame
}
\details{
Duplicate vars if grouping vars are also used by f
}
