% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_time_calculator.R
\name{obs_time_calculator}
\alias{obs_time_calculator}
\title{Calculate observation time}
\usage{
obs_time_calculator(focaltime, oos)
}
\arguments{
\item{focaltime}{data frame with duration of focal observation per focal animal}

\item{oos}{data frame with duration of out of sight time per focal animal}
}
\value{
a data frame
}
\description{
function calculates total focal time and total out of sight time from focaltime
then subtracts oos time from focal time
}
