% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_time_matrix.R
\name{obs_time_matrix}
\alias{obs_time_matrix}
\title{function that creates an obstime matrix}
\usage{
obs_time_matrix(obs_time, matrix_rows, matrix_cols)
}
\arguments{
\item{obs_time}{A data frame of total observation time
created by obs_time_calculator()}

\item{m_rows}{A character vector of IDs to go on row of matrix}

\item{m_cols}{A character vector of IDs to go on column of matrix}
}
\value{
A matrix with combined total observation time (in seconds)
for each dyad.
}
\description{
function that creates an obstime matrix
}
