% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prot_oos_duration_calculator.R
\name{prot_oos_duration_calculator}
\alias{prot_oos_duration_calculator}
\title{Calculate the duration of a behaviour without a partner.}
\usage{
prot_oos_duration_calculator(d_focal, start_codes, end_codes, focals)
}
\arguments{
\item{d_focal}{frame of focal observations.
Currently needs all these columns with exact same names:
"date", "protocol_start_time", "focal_animal", "observer", "action_time",
"actor", "action","at_sec_since_mdn"}

\item{start_codes}{a character vector of action codes
that indicate when a behaviour starts.}

\item{end_codes}{a character vector of action codes
that indicate when a behaviour ends.}

\item{focals}{a character vector of focal animals}
}
\value{
adds two columns to d_focal. One with the total duration in seconds
(duration_s), and another with the total duration in hh:mm:ss (duration)
}
\description{
Written specifically for focal protocol duration and out of sight time.
}
\examples{

}
