% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract_hms.R
\name{subtract_hms}
\alias{subtract_hms}
\title{Subtract two times in format "hh:mm:ss"}
\usage{
subtract_hms(time1, time2)
}
\arguments{
\item{time1}{time, will be subtracted by time2}

\item{time2}{time, will subtract from time 1}
}
\value{
Returns a character vector in format "hh:mm:ss"
}
\description{
Subtract two times in format "hh:mm:ss"
}
