% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_duration.R
\name{sum_duration}
\alias{sum_duration}
\title{Sum duration of grouped variables}
\usage{
sum_duration(df, ...)
}
\arguments{
\item{df}{a data.frame}

\item{...}{unquoted varialbel names to group_by}
}
\description{
returns the sum duration after grouping by variables
}
