% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_hms.R
\name{sum_hms}
\alias{sum_hms}
\title{Sum times in format "hh:mm:ss"}
\usage{
sum_hms(times)
}
\arguments{
\item{times}{a character vector of times}
}
\value{
returns a character vector of summed times
}
\description{
sums a character vector of times in format "hh:mm:ss"
}
