% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transp_sum_matrix.R
\name{transp_sum_matrix}
\alias{transp_sum_matrix}
\title{Transpose matrix and sum content}
\usage{
transp_sum_matrix(m)
}
\arguments{
\item{m}{A numerical matrix. Must be symetrical.}
}
\value{
A matrix
}
\description{
Transposes matrix and adds cell contents of transposed and original
}
